refactor: 코드 정리 및 구조 개선
- 주석 정리 및 ESLint 경고 해결
- 페이지 컴포넌트 폴더 구조 변경
