+++ src/basic/components/ui/Input.tsx
@@ -6,7 +6,6 @@ interface InputProps extends InputHTMLAttributes<HTMLInputElement> {
 }
 
 export default function Input({ label, className, ...props }: InputProps) {
-  // 공통 스타일
   const baseClasses =
     'w-full border-gray-300 rounded-md shadow-sm focus:ring-indigo-500 focus:border-indigo-500 px-3 py-2 border';
