+++ src/basic/models/cart.ts
@@ -1,20 +1,8 @@
 import { CartItem, Product } from '../../types';
 
 // 장바구니 비즈니스 로직 (순수 함수)
-// 모든 함수는 순수 함수로 구현 (부작용 없음, 같은 입력에 항상 같은 출력)
-//
-// 구현된 함수들:
-// 1. getMaxApplicableDiscount(item, cart): 적용 가능한 최대 할인율 계산
-// 2. calculateItemTotal(item, cart): 개별 아이템의 할인 적용 후 총액 계산
-// 3. getRemainingStock(product, cart): 남은 재고 계산
-// 4. calculateOriginalPrice(item): 장바구니 아이템의 원래 가격 계산
-//
-// 원칙:
-// - UI와 관련된 로직 없음
-// - 외부 상태에 의존하지 않음
-// - 모든 필요한 데이터는 파라미터로 전달받음
 
-// 적용 가능한 최대 할인율 계산 (순수 함수)
+// 적용 가능한 최대 할인율 계산
 export const getMaxApplicableDiscount = (item: CartItem, cart: CartItem[]): number => {
   const { discounts } = item.product;
   const { quantity } = item;
@@ -34,7 +22,7 @@ export const getMaxApplicableDiscount = (item: CartItem, cart: CartItem[]): numb
   return baseDiscount;
 };
 
-// 개별 아이템의 할인 적용 후 총액 계산 (순수 함수)
+// 개별 아이템의 할인 적용 후 총액 계산
 export const calculateItemTotal = (item: CartItem, cart: CartItem[]): number => {
   const { price } = item.product;
   const { quantity } = item;
@@ -43,7 +31,7 @@ export const calculateItemTotal = (item: CartItem, cart: CartItem[]): number => 
   return Math.round(price * quantity * (1 - discount));
 };
 
-// 남은 재고 계산 (순수 함수)
+// 남은 재고 계산
 export const getRemainingStock = (product: Product, cart: CartItem[]): number => {
   const cartItem = cart.find((item) => item.product.id === product.id);
   const remaining = product.stock - (cartItem?.quantity || 0);
@@ -51,7 +39,7 @@ export const getRemainingStock = (product: Product, cart: CartItem[]): number =>
   return remaining;
 };
 
-// 장바구니 아이템의 원래 가격 계산 (순수 함수)
+// 장바구니 아이템의 원래 가격 계산
 export const calculateOriginalPrice = (item: CartItem): number => {
   return item.product.price * item.quantity;
 };
